Loss_Time_Reporting/requirements.txt
arrow==0.14.2
Babel==2.8.0
Django==2.2
django-phone-field==1.8.0
djangorestframework==3.11.0
pkg-resources==0.0.0
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
